refactor, add views structure, employee sql, generated data
