+++ products.csv
@@ -1,0 +1,328 @@
+barcode,price,name,category,promoID,taxID
+1,$98.82 ,Branberries,food,1,5
+2,$44.42 ,Baby Donuts,food,2,3
+3,$42.04 ,Baby Sausage,food,3,2
+4,$43.05 ,Bacon Jerky,food,5,3
+5,$79.75 ,Bacon Sticks,food,1,3
+6,$90.61 ,BactoWeave,appliance,3,2
+7,$5.33 ,Baffledil,appliance,2,1
+8,$97.61 ,Bag It!,food,3,3
+9,$10.96 ,Bag of Scones,food,3,1
+10,$51.33 ,Balancing Action,appliance,2,5
+11,$68.65 ,Balkan Beef,food,1,3
+12,$85.46 ,Ball Lightning,service,2,3
+13,$40.19 ,Bamboo Gnu,material,5,3
+14,$20.27 ,Bamboolong,material,0,2
+15,$9.20 ,Bang’aBone,appliance,1,1
+16,$63.88 ,Burning Man Sunscreen,health,2,1
+17,$43.15 ,Barbie Real,toy,4,1
+18,$14.43 ,Barely Bacon,food,3,1
+19,$16.73 ,Barn Soup,food,3,4
+20,$73.11 ,Barnacle Power,service,0,2
+21,$10.82 ,Barnacle Tap,cleaning,2,4
+22,$79.37 ,Barney Toy,service,4,5
+23,$19.20 ,BarTelevision,entertainment,1,2
+24,$22.56 ,Basil Bromide,cleaning,0,5
+25,$51.51 ,Bath and Relax,service,5,1
+26,$12.36 ,Bathe Away,service,2,3
+27,$89.90 ,Bath-Lite,appliance,4,5
+28,$6.12 ,Battle Screen,entertainment,2,1
+29,$28.33 ,Baubles,cleaning,4,2
+30,$17.37 ,BBQ’d Napkins,food,5,1
+31,$99.36 ,Bead2Bead,service,3,5
+32,$82.28 ,Beadniks,food,3,2
+33,$40.38 ,BeanCounter,material,4,3
+34,$55.80 ,BeanThere,service,4,1
+35,$15.47 ,Beauty in a Bottle,health,0,2
+36,$63.13 ,BeautyStretch,health,5,5
+37,$16.28 ,Bedder,service,0,2
+38,$10.12 ,Bed Springs,furniture,3,2
+39,$23.22 ,Bedleham Trip,service,3,2
+40,$86.87 ,Bedware,furniture,0,4
+41,$30.58 ,Bee Salt,food,2,4
+42,$84.93 ,Bee Salt Cream,food,0,1
+43,$88.28 ,Bee Saw,appliance,4,2
+44,$26.62 ,Beef Blackout,food,1,2
+45,$85.00 ,Beef Blanks,food,0,3
+46,$34.58 ,Beef Bowl Knoll,food,5,3
+47,$38.13 ,Beefbelly,food,0,1
+48,$96.74 ,Beer Puppets,food,4,4
+49,$71.17 ,Beeragra,service,0,1
+50,$29.36 ,Beegle Care,service,1,4
+51,$61.04 ,Belt,clothing,2,1
+52,$40.36 ,Beige and Briney,service,4,1
+53,$10.44 ,Bela Bites,food,0,2
+54,$46.85 ,Bella Outdoors,service,2,5
+55,$21.14 ,Belly Jelly,health,2,5
+56,$35.36 ,Belly-flop cream,health,0,3
+57,$57.64 ,Beluga Gum,food,0,4
+58,$47.74 ,Bermo,food,4,2
+59,$5.77 ,Berryboy,food,0,5
+60,$76.64 ,Besto,sed,1,4
+61,$29.35 ,Better Then Chicken,food,1,2
+62,$5.56 ,Bewildering Chicken,food,3,1
+63,$41.49 ,Beyond Bedding,furniture,3,1
+64,$36.94 ,BeyondSleep,furniture,2,1
+65,$45.06 ,Billy Boy Pants,clothing,1,5
+66,$4.50 ,Bicyclone,entertainment,3,5
+67,$88.26 ,BiffPaste,health,3,2
+68,$57.82 ,Biflex,health,3,1
+69,$21.90 ,Big Juicy Apple,food,5,1
+70,$36.39 ,BigCookie,food,5,1
+71,$76.35 ,BigTent,entertainment,2,4
+72,$51.46 ,BigWay,entertainment,5,2
+73,$18.19 ,Bike Along,service,5,1
+74,$65.95 ,Billion Air,service,5,1
+75,$51.81 ,Bimbowl,appliance,3,1
+76,$69.96 ,Binarium,entertainment,1,1
+77,$21.82 ,Bingles,food,4,5
+78,$78.74 ,Bingo Storm,service,2,5
+79,$79.35 ,Biolid,health,0,4
+80,$60.32 ,Biological Interface,appliance,5,3
+81,$41.00 ,Birthday Potato,food,0,5
+82,$35.71 ,Biscotti de Chien,food,2,2
+83,$85.73 ,Bite’Ems,food,5,3
+84,$86.04 ,Bitskater,service,0,3
+85,$34.92 ,Bitter Plaid,clothing,1,5
+86,$25.77 ,Bittor,entertainment,3,4
+87,$61.66 ,Bizmatic,appliance,0,1
+88,$53.59 ,BizSweeper,appliance,3,1
+89,$70.34 ,BizWhiz,appliance,1,5
+90,$20.67 ,Black Belt Barbie,toy,5,1
+91,$24.90 ,Black Vinyl,material,2,5
+92,$81.09 ,Blahg,material,0,3
+93,$38.76 ,Blanked Beef,food,4,5
+94,$15.63 ,Blazarus Trip,service,4,3
+95,$50.30 ,Blaze Chicken,food,1,5
+96,$78.46 ,Bleach Melba,cleaning,0,2
+97,$31.46 ,Blind Truss,service,1,1
+98,$69.25 ,Blink!,entertainment,2,2
+99,$9.34 ,Blink! 2,entertainment,0,5
+100,$20.76 ,Blister Burger,food,2,1
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
+,,,,,
