--- ryanwwest.md
@@ -1,5 +1,0 @@
-I'd love to join this project as I'm interested in inventory management. I eventually want to build a personal inventory management system and feel like this project would be good preparation.
-
-I also have work experience with the entire tech stack. I've used Python for a few years and am currently building a DBMS client for MongoDB at Qualtrics.
-
-I want to work hard and complete this project quickly.
