+++ view/developer.py
@@ -1,48 +1,18 @@
-import psycopg2
-from psycopg2 import OperationalError
-
 import tkinter as tk
-from tkinter import messagebox
-
-
-def create_connection(db_name, db_user, db_password, db_host, db_port):
-    connection = None
-    try:
-        connection = psycopg2.connect(
-            database=db_name,
-            user=db_user,
-            password=db_password,
-            host=db_host,
-            port=db_port,
-        )
-        print("Connection to PostgreSQL DB successful")
-    except OperationalError as e:
-        print(f"The error '{e}' occurred")
-    return connection
-
-def execute_query(connection, query):
-    connection.autocommit = True
-    cursor = connection.cursor()
-    try:
-        cursor.execute(query)
-        print("Query executed successfully")
-    except OperationalError as e:
-        print(f"The error '{e}' occurred")
-
-
-
+from db import postgres
 
 def create_products():
     create_products_table = """
         CREATE TABLE IF NOT EXISTS products (
         barcode INTEGER PRIMARY KEY,
+        price NUMERIC(2),
         name TEXT NOT NULL, 
         category TEXT,
         promoID INTEGER,
         taxID INTEGER
         )
         """
-    execute_query(connection, create_products_table)
+    postgres.execute_query(create_products_table)
 
 def create_taxes():
     create_taxes_table = """
@@ -52,7 +22,7 @@ def create_taxes():
         rate INTEGER
         )
         """
-    execute_query(connection, create_taxes_table)
+    postgres.execute_query(create_taxes_table)
 
 def create_promos():
     create_promos_table = """
@@ -62,31 +32,28 @@ def create_promos():
         discount INTEGER
         )
         """
-    execute_query(connection, create_promos_table)
+    postgres.execute_query(create_promos_table)
 
 
 def drop_products():
     drop_products_table = """ 
         DROP TABLE IF EXISTS products;
         """
-    execute_query(connection, drop_products_table)
+    postgres.execute_query(drop_products_table)
 
 def drop_taxes():
     drop_products_table = """ 
         DROP TABLE IF EXISTS taxes;
         """
-    execute_query(connection, drop_products_table)
+    postgres.execute_query(drop_products_table)
 
 def drop_promos():
     drop_products_table = """ 
         DROP TABLE IF EXISTS promos;
         """
-    execute_query(connection, drop_products_table)    
+    postgres.execute_query(drop_products_table)
 
 
-connection = create_connection(
-    "postgres", "postgres", "mysecretpassword", "localhost", "5432"
-)
 
 window = tk.Tk()
 window.title("create and drop tables")
@@ -113,5 +80,3 @@ drop_promos_button = tk.Button(text="drop table promos", command=drop_promos)
 drop_promos_button.pack()
 
 window.mainloop()
-
-
