+++ view/manager.py
@@ -1,0 +1,3 @@
+import tkinter as tk
+from db import postgres
+
